backend/requirements.txt
msgraph-sdk
python-dateutil
python-dotenv
azure-ai-formrecognizer
pytz
autogen-agentchat
autogen-ext[azure]
autogen-ext[openai]
fastapi
uvicorn